Use atoi instead of atol

+++ c/fibo_strings.c
@@ -106,7 +106,7 @@ int main(int argc, char* argv[]) {
     int n = 4784969;               // The first Fibonacci number with a million digits
 
     if (argc >= 2) {
-        n = atol(argv[1]);
+        n = atoi(argv[1]);
     }
 
     init_memo();
